refractor: move semantic `add_builtin` into `builtins.c`
- refractor: move `Scope`, `Symbol`, `Sema` to `sema.h`
- refractor: add `Sema` prefix in `Scope`, `Symbol` typedef (`SemaCtx` for `Sema`)
- refractor: move semantic scope management to `scope.c`
- refractor: move semantic analysis to `Sema/`
